refactor(grid): add tooltip plugin
