+++ packages/vue/src/grid/src/sort/src/methods.ts
@@ -1,0 +1,1 @@
+export default {}
