+++ packages/vue/src/grid/src/table/src/events.ts
@@ -242,40 +242,3 @@ export function handleGlobalKeydownEvent(event) {
 export function handleGlobalResizeEvent() {
   this.recalculate()
 }
-
-// 触发表头 tooltip 事件
-export function triggerHeaderTooltipEvent(event, params) {
-  let { tooltipStore } = this
-  let { column, showHeaderTip } = params
-  if (tooltipStore.column !== column || !tooltipStore.visible) {
-    // 在 v3.0 中废弃 label
-    this.handleTooltip(event, column, null, showHeaderTip, true)
-  }
-}
-
-// 触发表尾 tooltip 事件
-export function triggerFooterTooltipEvent(event, params) {
-  let { column } = params
-  let tooltipStore = this.tooltipStore
-  if (tooltipStore.column !== column || !tooltipStore.visible) {
-    this.handleTooltip(event, column)
-  }
-}
-
-// 触发 tooltip 事件
-export function triggerTooltipEvent(event, params) {
-  let { editConfig, editStore, tooltipStore } = this
-  let { actived } = editStore
-  let { row, column, showTip } = params
-  if (editConfig) {
-    if (
-      (editConfig.mode === 'row' && actived.row === row && column.editor) ||
-      (actived.row === row && actived.column === column)
-    ) {
-      return
-    }
-  }
-  if (tooltipStore.column !== column || tooltipStore.row !== row || !tooltipStore.visible) {
-    this.handleTooltip(event, column, row, showTip)
-  }
-}
